+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/dev-launchpad
 
+## 2.8.51
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.50
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.49
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.48
+
 ## 2.8.47
 
 ### Patch Changes