+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @stackframe/example-demo-app
 
+## 2.8.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.51
+  - @stackframe/stack@2.8.51
+  - @stackframe/stack-ui@2.8.51
+
+## 2.8.50
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.50
+  - @stackframe/stack-shared@2.8.50
+  - @stackframe/stack-ui@2.8.50
+
+## 2.8.49
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.49
+  - @stackframe/stack-shared@2.8.49
+  - @stackframe/stack-ui@2.8.49
+
+## 2.8.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.48
+  - @stackframe/stack@2.8.48
+  - @stackframe/stack-ui@2.8.48
+
 ## 2.8.47
 
 ### Patch Changes